refactor(backend): move thrift/go/lex_yacc rule groups to target modules (M2 of #1264)
Second M2 step: migrate the remaining non-cc-family rule groups out of
backend.py, relocating their small helpers along with them.

- thrift -> thrift_library.py: provider + the (thrift-only) _incs_list_to_string
  helper moved out of backend.py.
- go -> go_targets.py: provider; emits the conditional golang_pool line via
  ctx.add_line and is a no-op unless go is configured (unchanged behavior).
- lex_yacc -> lex_yacc_target.py: provider + the _find_win_bison_data_dir
  helper (now a module function); needs glob/shutil/os which the module now
  imports.
- backend.py: delete the three generate_*_rules methods, _find_win_bison_data_dir,
  _incs_list_to_string, the three delegating lambdas, and the now-unused
  glob/shutil imports.

Verified: build.ninja rule section byte-identical to baseline on blade-test
(MSVC); registry unit tests green; real lex_yacc_basic build passes.

Remaining in backend.py for dedicated PRs: the cc family (cc, cuda), java_scala,
proto, version.
